% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.annot.R
\docType{data}
\name{gene.annot}
\alias{gene.annot}
\title{Gene annotation data (hg38)}
\format{A matrix with 26417 rows and 3 columns
\describe{
\item{chr}{chromosome containing the gene of interest}
\item{pos}{genomic position (base pairs) for the gene of interest}
\item{cytoband}{cytoband containing the gene of interest}
\item{Row names}{gene symbols}
}}
\source{
\url{https://usegalaxy.org/}
}
\usage{
gene.annot
}
\description{
Genomic position and cytoband annotation data.
}
\keyword{datasets}
