% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.annot.R
\docType{data}
\name{sample.annot}
\alias{sample.annot}
\title{Sample annotation data}
\format{A matrix with 26417 rows and 3 columns
\describe{
\item{Barcode}{sample identifier}
\item{New.HPV.Status}{HPV infection status}
}}
\source{
\url{http://www.nature.com/nature/journal/v517/n7536/full/nature14129.html}
}
\usage{
sample.annot
}
\description{
Human papillomavirus (HPV) infection status for the n = 279 patients with head and neck
squamous cell carcinoma in the The Cancer Genoma Atlas cohort.
}
\keyword{datasets}
