% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.exp.convert.R
\name{tcga.exp.convert}
\alias{tcga.exp.convert}
\title{A Function for Reformatting TCGA mRNA Expression Matrices}
\usage{
tcga.exp.convert(exp.mat)
}
\arguments{
\item{exp.mat}{A matrix of mRNA expression data.  Typically illuminahiseq_rnaseqv2-RSEM_genes_normalized, or a subset thereof,
including the header rows.}
}
\value{
A matrix of mRNA expression data (rows = genes, columns = samples) that is suitable for input to mVisAGe functions.
}
\description{
This function reformats mRNA expression matrices obtained from the Broad Institute's Firehose GDAC 
 (\url{https://gdac.broadinstitute.org/}) so they can be used as input for mVisAGe functions.
}
\examples{
exp.mat = tcga.exp.convert(exp.mat)

}
